remove fromatData fun to reperate js file
